@@ -1,21 +1,13 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { AppContext } from './AppContext';
 import App from './App';
 
-// const element = document.getElementById('fluent-translate') as HTMLElement & {
-//   shadowRoot: ShadowRoot;
-// };
-// element.attachShadow({ mode: 'open' });
-
-// const style = document.createElement('style');
-// style.innerHTML = stitchesConfig.getCssText();
-// element.shadowRoot.appendChild(style);
-
-ReactDOM.render(
+const container = document.getElementById('fluent-translate');
+const root = createRoot(container!);
+root.render(
   <AppContext>
     <App />
-  </AppContext>,
-  document.getElementById('fluent-translate')
+  </AppContext>
 );
