@@ -19,18 +19,23 @@ export const Select = styled('select', {
   'bc': 'transparent',
   'backgroundImage':
     'url(\'data:image/svg+xml;utf-8,<svg width="10" height="10" viewBox="0 0 16 12" xmlns="http://www.w3.org/2000/svg"><path stroke-width="2" d="M11.3613 2.73633L11.8887 3.26367L6 9.15234L0.111328 3.26367L0.638672 2.73633L6 8.09766L11.3613 2.73633Z"></path></svg>\')',
-  '&:hover': {
-    bc: '$flatHover',
+  '@hover': {
+    '&:where(:hover, :focus)': {
+      bc: '$surface2',
+    },
+    '&:where(:focus)': {
+      borderColor: '$surface3',
+    },
   },
-  '&:focus': {
-    bc: '$flatActive',
-    borderColor: '$borderActive',
+  '&:where(:active, [data-state="open"])': {
+    bc: '$surface3',
+    borderColor: '$surface4',
   },
   '& optgroup': {
-    bc: '$surfaceStill',
+    bc: '$background',
   },
 
-  '@media (prefers-color-scheme: dark)': {
+  '@dark': {
     backgroundImage:
       'url(\'data:image/svg+xml;utf-8,<svg width="10" height="10" viewBox="0 0 16 12"  xmlns="http://www.w3.org/2000/svg"><path stroke="rgb(255,255,255)" stroke-width="2" d="M11.3613 2.73633L11.8887 3.26367L6 9.15234L0.111328 3.26367L0.638672 2.73633L6 8.09766L11.3613 2.73633Z"></path></svg>\')',
   },
